backend/requirements.txt
groq
pydantic>=2.0.0
apify-client
requests
python-dotenv
fastapi
uvicorn
openai
twilio
gspread
google-auth
gradio
